Refactor error handling and configuration management in clip-sync-repair
- Improved error handling in `locate_query.rs` by replacing nested `if` statements with `match` expressions, enhancing readability and maintainability.
- Updated `rubato.rs` to clear `decoded_sample_count` after resampling to prevent unit mixing, ensuring accurate state management.
- Enhanced `scan_gaps.rs` to log warnings on B-side silence scan failures, improving observability during gap scanning.
- Introduced `profile_field_mask` in `RepairConfig` to preserve explicit TOML settings across CLI overrides, ensuring user configurations are respected.
- Added validation checks in `RepairConfig` to reject non-finite float values, enhancing robustness against invalid configurations.
- Refactored `ffmpeg_mux.rs` to streamline duration validation and improve error handling during muxing operations, ensuring better resource management.

+++ crates/clip-sync-repair-harness/src/dual_fit_oracle.rs
@@ -301,7 +301,7 @@ fn validate_stepped_oracle(built: &BuiltFloorOracle, defaults: &FloorOracleDefau
         .map(|s| s.unsigned_abs())
         .max()
         .unwrap_or(0);
-    if peak > built.meta.gap_interior_peak_max as u16 {
+    if peak > built.meta.gap_interior_peak_max {
         return Err(format!(
             "gap interior peak {peak} exceeds max {} after encode",
             built.meta.gap_interior_peak_max
